fix - force return of int or long column types. Don't leave it up to database to decide.

+++ src/main/java/org/knowm/yank/Yank.java
@@ -566,7 +566,7 @@ public class Yank {
       ColumnListHandler<T> resultSetHandler;
       if (columnType.equals(Integer.class)) {
         resultSetHandler = (ColumnListHandler<T>) new IntegerColumnListHandler(columnName);
-      } else if (columnType.equals(Integer.class)) {
+      } else if (columnType.equals(Long.class)) {
         resultSetHandler = (ColumnListHandler<T>) new LongColumnListHandler(columnName);
       } else {
         resultSetHandler = new ColumnListHandler<T>(columnName);
